requirements.txt
django
gunicorn
djangorestframework
markdown
pillow
django-filter
whitenoise
django-heroku